Два junit-теста, собранные в один test suite
