+++ 11/.gitignore
@@ -1,3 +1,1 @@
-/Solution.class
-*.classpath
-*.project
+.settings/
