+++ 11/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>solution1</name>
+	<name>COP_solutions</name>
 	<comment></comment>
 	<projects>
 	</projects>
