+++ 11/first_package/.gitignore
@@ -1,0 +1,3 @@
+*.class
+*.classpath
+*.project
