+++ 11/first_package/AllTests.java
@@ -1,0 +1,11 @@
+package first_package;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+@RunWith(Suite.class)
+@SuiteClasses({ FirstCase.class })
+public class AllTests {
+
+}
