+++ 11/first_package/MainClass.java
@@ -1,0 +1,72 @@
+package first_package;
+
+import java.io.*;
+import java.util.*;
+import java.util.Arrays;
+
+public class MainClass {
+	/*
+	    * Определить максимальную мощность подмножества элементов заданного массива. Элементы включаются в подмножество, если
+		* разность между любыми элементами подмножества меньше либо равна 1.	
+		* Дано: n - количество чисел в массиве
+		*		a - массив чисел
+		* 		2 <= n <= 100
+		*		0 < a[i] < 100
+		*		ответ должен быть >=2, задаются такие данные
+		* Пример:
+		*  9
+		*  1 1 2 2 4 4 5 5 5
+		* Ответ:
+		* 5
+		*
+		* 6
+		* 4 6 5 3 3 1
+		* Ответ:
+		* 3
+		*/
+		
+	    static int func1(int[] A) 
+	    {
+	    	int n = A.length;
+	    	Arrays.sort(A);
+	    	int max = 0;
+	        int s = 0;       
+	        
+	        for(int i = 0; i < n; i++) 
+	        {
+	            for(int j = i; j < n; j++)
+	                if(A[j] - A[i] <= 1) 
+	                    s++;
+	   
+	            if(s > max)
+	                max = s;
+	            s = 0;
+	        }
+	        
+	        return max;
+	      }
+	    
+
+	    private static final Scanner scanner = new Scanner(System.in);
+
+	    public static void main(String[] args) throws IOException {
+
+	    	
+	        int n = scanner.nextInt();
+	        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");
+
+	        int[] a = new int[n];
+
+	        String[] aItems = scanner.nextLine().split(" ");
+	        scanner.skip("(\r\n|[\n\r\u2028\u2029\u0085])?");
+
+	        for (int i = 0; i < n; i++) {
+	            int aItem = Integer.parseInt(aItems[i]);
+	            a[i] = aItem;
+	        }
+
+	        System.out.print(func1(a));;		
+
+	        scanner.close();
+	    }
+}
